chore: refactor project structure and update main entry points
- Updated main entry in package.json to point to the new directory structure.
- Added project structure section to README for better clarity.
- Adjusted tsconfig.json to reflect new root directory and include templates.
- Updated wrangler.toml to align with the new file paths.
- Removed outdated Cloudflare-specific files and directories to streamline the codebase.
