+++ packages/api/index.ts
@@ -1,0 +1,2 @@
+// API exports
+export * from './routes/configPage';
