+++ packages/core/index.ts
@@ -1,0 +1,4 @@
+// Core modules export
+export * from './config/configManager';
+export * from './catalog/catalogAggregator';
+export * from './utils/manifestBuilder';
